CMake: 3.17.3+
We don't use the DEVICE_LINK features in upstream AMReX logic yet,
so we can stay with 3.17.3+ for now.

For CMake < 3.18.0 we need to overspecify the Python dependencies
that we are actually looking for. This is mainly a problem with
`manylinux` images (but there we can upgrade CMake with `pip`
first) and fine-grained package managers.

Generally, anticipate a switch of AMReX as a whole to CMake 3.19+
in the mid-term because a lot of niceness landed in it.