Clean Pip Staging Directory
he distutils staging (`build` -> `_tmppythonbuild`) directory that
pip uses to collect build artifacts is not by default cleaned between
multiple `pip` runs.

This can be confusing when we recompile, because old `amrex_pybind*` files
can be still in it that are not present in our own `build/lib/`.

This cleans that staging directory before builds now.
It also sets the whole `build_base` so that no artifact lands in the
default, which was `build/`.