MultiFab: Fix Fixture Lifetime (#84)
pytest fixtures are cached by default. That's not ideal for us,
since we want to clean up and re-init AMReX and its memory
arenas between tests.

Thus, we now create a multifab fixture that is just a generator,
so that the returned/yielded object can be cached - but the
generated MultiFab is actually destroyed as we would expect once
the test finishes.