Doc: Multi-Dim and Python Executable
Documents how to overwrite the SPACEDIM default of AMReX from "3"
to building 1, 2 and 3 dimensional
versions at the same time.
Example: `-DAMReX_SPACEDIM="1;2;3"`

Documents how to overwrite the automatically found Python executable,
which is useful if one is picked that is undesired. Typically happens
outside of virtual environments and on macOS.
Example: `-DPython_EXECUTABLE=$(which python3)`