Install: CMake Config & Targets
Install the CMake package files for consumption of CMake targets
in downstream projects.

We will in the future expose public headers for downstream projects
as well that will help us to compile templated classes, such as
particle containers, targeted in the downstream repo instead of
collecting all known usage here in upstream.