For MuliFab indexing, () means all cells in all dimension, valid and ghost (#438)
This PR clarifies the use of the empty tuple index, `()`, for MultiFabs.
Giving the empty tuple as the only index of a MultiFab will return the
full array, both valid and ghost cells. This is in contrast to the
`Ellipsis` that will return the array with only value cells.
For example
```
mf = MultiFab()
mf[()]  # returns valid and ghost cells
mf[...]  # returns only valid cells
```

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>