Fix Use-After-Free in NumPy/df (#478)
For AoS and SoA `.to_numpy()` and dependent logic
like `.to_df()`, the lifetime of the temporary copied variable is only
handled by NumPy if we first create a named variable.

The `to_host()` returned object is a temporary, and `np.array` using the
`__array_interface__` protocol does not keep it alive automatically
unless it is stored in an actual variable (eg., `tmp`).

X-ref:
-
https://github.com/conda-forge/impactx-feedstock/pull/56#issuecomment-3304382067
-
https://github.com/AMReX-Codes/pyamrex/pull/458#issuecomment-3304399453